+++ @jambaree/next-wordpress/index.ts
@@ -6,9 +6,13 @@ import FlexibleContent from "./src/flexibleContent";
 import useMenuItems from "./src/useMenuItems";
 import getData from "./src/getData";
 import getYoastData from "./src/getYoastData";
+import { generateStaticParams } from "./src/generateStaticParams";
+import { generateMetadata } from "./src/generateMetadata";
 
 export {
   getYoastData,
+  generateStaticParams,
+  generateMetadata,
   getData,
   getSeedData,
   getTemplate,